Move add_directory up a level
